requirements.txt
stix2==3.0.1
taxii2-client==2.3.0
requests==2.31.0
click==8.1.7
pyyaml==6.0.1
python-dateutil==2.8.2
tabulate==0.9.0
colorama==0.4.6
